requirements.txt
Django==4.0.4
django_extensions==3.1.5
djangorestframework==3.13.1
narvan-commons-lib==3.2.8
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
kavenegar==1.1.2
narvan-types-lib==1.0.36
bazdeh-server-commons-lib==1.27.9
narvan-grpc-lib==1.1.23
redis==3.5.3
drf-spectacular==0.22.1
psycopg2-binary==2.9.3
grpcio==1.46.3
protobuf==3.19.1
django-prometheus==2.2.0
gunicorn==20.1.0
elasticsearch==8.3.1
ordered-set==4.1.0
sentry-sdk==1.8.0
marshmallow==3.17.0
djangorestframework-recursive==0.1.2
python-dateutil==2.8.2
django-workers==0.2.0
prometheus-client==0.14.1
djangogrpcframework==0.2.1
requests==2.28.1
Pillow==9.2.0
python-bidi==0.4.2
arabic-reshaper==2.1.4
